added a new Setting Screen page with the go back arrow
