+++ src/Components/Homepage.jsx
@@ -55,7 +55,7 @@ export default class Homepage extends React.Component {
         if (this.state.lowerPanelContent === 'clinicInfo') this.setLowerPanelContent('findCare');
         if (this.state.lowerPanelContent === 'learn') this.setLowerPanelContent('selection');
         if (this.state.lowerPanelContent === 'tips') this.setLowerPanelContent('selection');
-        if (this.state.lowerPanelContent === 'edit') this.setLowerPanelContent('selection');
+        
     };
 
 
@@ -81,7 +81,9 @@ export default class Homepage extends React.Component {
                             lowerPanelContent={this.state.lowerPanelContent}
                             goBack={this.goBack}
                             setLowerPanelContent={this.setLowerPanelContent}
-                            getLocalizedText={this.props.getLocalizedText}/>
+                            getLocalizedText={this.props.getLocalizedText}
+                            setAppState={this.props.setAppState}
+                            />
             </View>
         )
     }
